Fixes #19602 - Disable auth login when sessions are off

+++ lib/hammer_cli_foreman/auth.rb
@@ -10,6 +10,11 @@ module HammerCLIForeman
       option ["-p", "--password"], "PASSWORD", _("password to access the remote system")
 
       def execute
+        if !(HammerCLIForeman.foreman_api_connection.authenticator.is_a?(HammerCLIForeman::Api::SessionAuthenticatorWrapper))
+          print_message(_("Can't perform login. Make sure sessions are enabled in hammer configuration file."))
+          return HammerCLI::EX_USAGE
+        end
+
         HammerCLIForeman.foreman_api_connection.logout
         context[:api_connection].drop_all
         HammerCLI::Settings.load({
